refactor(constant): 패키지 정리
수정 내용
- 가독성을 높이기 위해 input_output_message를
  view_message로 변경
- Separator 열거형을 InputSeparator로 변경한 후
  constant.input 패키지로 옮김
