+++ src/main/java/store/constant/view_message/InputExceptionMessage.java
@@ -1,4 +1,4 @@
-package store.constant.input_output_message;
+package store.constant.view_message;
 
 public enum InputExceptionMessage {
     INPUT_IS_WRONG_FORMAT("올바르지 않은 형식으로 입력했습니다. 다시 입력해 주세요."),
